apps/api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mustory-api"
version = "0.1.0"
description = "FastAPI backend for the Mustory MVP."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<0.116",
  "uvicorn[standard]>=0.30,<0.31",
  "pydantic[email]>=2.9,<2.10",
  "sqlalchemy>=2.0,<2.1",
  "alembic>=1.13,<1.14",
  "psycopg[binary]>=3.2,<3.3",
  "python-dotenv>=1.0,<2.0",
  "redis>=5.0,<5.2",
  "rq>=1.16,<2.0",
  "httpx>=0.27,<0.28",
  "supabase>=2.0,<3.0",
  "boto3>=1.34,<2.0",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.11,<1.12",
  "ruff>=0.6.1,<0.7",
  "pytest>=8.3,<8.4",
  "pytest-asyncio>=0.23,<0.24",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
plugins = ["pydantic.mypy"]
strict = true

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.hatch.build.targets.sdist]
include = ["app", "README.md", "pyproject.toml"]
